requirements.txt
altgraph==0.17
arabic-reshaper==2.1.3
click==7.1.2
colorama==0.4.4
cycler==0.10.0
et-xmlfile==1.1.0
fpdf==1.7.2
future==0.18.2
html5lib==1.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
pdfkit==0.6.1
pefile==2021.5.24
Pillow==8.3.1
plotly==5.1.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
pywin32-ctypes==0.2.0
reportlab==3.5.68
scipy==1.7.1
six==1.16.0
tenacity==8.0.1
typer==0.3.2
webencodings==0.5.1
xhtml2pdf==0.2.5
